Move PRBS functions to own module
